components/easydoc-backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
requests==2.31.0
python-dotenv==1.0.1
PyMuPDF==1.23.6
google-generativeai==0.4.1
